requirements.txt
Flask==1.0.2
py-healthcheck
requests
tox==2.5.0
functools